Refactoring the package
