+++ nmt_chainer/models/rnn_cells.py
